Added better behavior if a protocol isn't supported.
ColdJava now ships with two protocols built in: the http protocol it used to, an a NotSupported protocol which just returns "This protocol is not supported". The ProtocolHandler, if it detects an HTTP code 501 after the CLASS request, returns an instance of this protocol instead. Additionally, the crucial line that caches the protocols was actually added because reasons.